Fixed DASM on x86_64 and added ELIZA-esque sample in samples/eliza.asm

+++ src/dasm/ast.d
@@ -731,7 +731,7 @@ alias Callable!(Type.Builtin|Type.Keyword)              BuiltinKeyword; ///Synta
 
 class Scope : Expression {
     Expression[] defines;       ///Object definitions.
-    uint[string] symbols;       ///Other symbols.
+    size_t[string] symbols;       ///Other symbols.
     Scope outter;               ///Outter scope.
 
     this(Scope outter = null, uint line = __LINE__, string file = __FILE__) {
